tts_pipeline/requirements.txt
TTS==0.22.0
faster-whisper
pydub
librosa
soundfile
numba
numpy
scipy
python-dotenv
PyYAML
torch
torchaudio